refactor: improve imports by harnessing path aliases and barrel exports
