+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useContent } from "./useContent";
+export { useImages } from "./useImages";
